MonolithicSystem/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
python-jose
minio
psycopg2-binary  
